working on fsm edit system 5 delete asks for confirm if children of node exist

+++ front-end/react/pages/morgueFoldersPage.js
@@ -83,8 +83,28 @@ export default class MorgueFoldersPage extends Component {
   deleteFolder(e)
   {
       e.preventDefault();
-      let newFolderData = storageService.deleteFolder(this.state.selectedKey)
-      this.setState({ selectedKey: "",folderData: newFolderData,selectedFolderName: "", actionMode:"INIT"})
+      let idxObj = storageService.getIndex(); 
+      let folderData = idxObj[this.state.selectedKey]
+      if (folderData.children.length > 0)
+      {
+          var retVal = confirm("WARNING: Deleting will remove all children and contents!\n Do you want to continue?");
+          if (retVal === true)
+          {
+               let newFolderData = storageService.deleteFolder(this.state.selectedKey)
+               this.setState({ selectedKey: "",folderData: newFolderData,selectedFolderName: "", actionMode:"INIT"})
+          }
+          
+          
+      }
+      else
+      {
+          
+          let newFolderData = storageService.deleteFolder(this.state.selectedKey)
+          this.setState({ selectedKey: "",folderData: newFolderData,selectedFolderName: "", actionMode:"INIT"})
+          
+          
+      }
+      
       
   }
  
@@ -156,7 +176,7 @@ export default class MorgueFoldersPage extends Component {
       if (type === 'DELETE')
       {
           css = "btn btn-red space-left"
-          if (this.state.selectedKey === "/0")
+          if (this.state.selectedKey === "/0" || this.state.actionMode != 'CHOOSE')
           {
              css = css + " hidden"
           }
